+++ optistim/optimization/fes_multi_start.py
@@ -4,9 +4,11 @@ import numpy as np
 import pickle
 
 from bioptim import Solver, MultiStart, Solution
-from .model import DingModelFrequency, DingModelPulseDurationFrequency, DingModelIntensityFrequency
+from optistim import (DingModelFrequency,
+                      DingModelPulseDurationFrequency,
+                      DingModelIntensityFrequency)
 from .fes_ocp import FunctionalElectricStimulationOptimalControlProgram
-from .read_data import ExtractData
+from ..read_data import ExtractData
 
 
 class FunctionalElectricStimulationMultiStart(MultiStart):
